backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
cloudinary==1.37.0
anthropic==0.18.0
schedule==1.2.0
APScheduler==3.10.4
croniter==2.0.2
psutil==7.2.2
pyppeteer==1.0.2
